Plugin signature updated
