+++ .gitignore
@@ -5,3 +5,4 @@ tmp/*
 dist/*
 
 coverage.out
+.vscode/*
