+++ go.mod
@@ -3,14 +3,9 @@ module github.com/gadget-bot/gadget
 go 1.16
 
 require (
-	astuart.co/goq v1.0.0
-	github.com/coreos/go-systemd v0.0.0-20190321100706-95778dfbb74e // indirect
-	github.com/go-sql-driver/mysql v1.6.0
-	github.com/likexian/whois v1.12.1
 	github.com/rs/zerolog v1.23.0
 	github.com/slack-go/slack v0.9.1
-	golang.org/x/lint v0.0.0-20210508222113-6edffad5e616 // indirect
-	golang.org/x/tools v0.1.2 // indirect
+	github.com/stretchr/testify v1.3.0 // indirect
 	gorm.io/driver/mysql v1.1.1
 	gorm.io/gorm v1.21.11
 )
